+++ packages/callwave_flutter_method_channel/android/src/main/kotlin/com/callwave/flutter/methodchannel/manager/CallNotificationManager.kt
@@ -193,6 +193,10 @@ class CallNotificationManager(
         }
     }
 
+    fun hideIncomingUi(callId: String) {
+        dismissIncoming(callId, stopForegroundService = false)
+    }
+
     fun dismissMissed(callId: String) {
         notificationManagerCompat.cancel(missedNotificationId(callId))
     }